modified

+++ controllers/categoryController.js
@@ -5,7 +5,7 @@ const generateCategoryId = async () => {
   return `category${String(count + 1).padStart(4, "0")}`;
 };
 
-const createMainCategory = async (req, res) => {
+exports.createMainCategory = async (req, res) => {
   try {
     const { mainCategory } = req.body;
 
@@ -26,7 +26,7 @@ const createMainCategory = async (req, res) => {
   }
 };
 
-const addSubCategory = async (req, res) => {
+exports.addSubCategory = async (req, res) => {
   try {
     const { mainCategory, subCategoryName } = req.body;
 
@@ -49,7 +49,7 @@ const addSubCategory = async (req, res) => {
   }
 };
 
-const addSubSubCategory = async (req, res) => {
+exports.addSubSubCategory = async (req, res) => {
   try {
     const { mainCategory, subCategoryName, subSubCategoryName } = req.body;
 
@@ -76,7 +76,96 @@ const addSubSubCategory = async (req, res) => {
   }
 };
 
-const getCategories = async (req, res) => {
+exports.addLavel3Category = async (req, res) => {
+  try {
+    const {
+      mainCategory,
+      subCategoryName,
+      subSubCategoryName,
+      lavel3CategoryName,
+    } = req.body;
+
+    const category = await Category.findOne({ mainCategory });
+    if (!category) {
+      return res.status(404).json({ message: "Main category not found" });
+    }
+
+    const subCategory = category.subcategories.find(
+      (sub) => sub.name === subCategoryName
+    );
+    if (!subCategory) {
+      return res.status(404).json({ message: "Subcategory not found" });
+    }
+
+    const subSubCategory = subCategory.subsubcategories.find(
+      (subSub) => subSub.name === subSubCategoryName
+    );
+    if (!subSubCategory) {
+      return res.status(404).json({ message: "Subsubcategory not found" });
+    }
+
+    subSubCategory.lavel3CategorySchema.push({ name: lavel3CategoryName });
+    await category.save();
+    res.status(201).json({
+      message: "3rd level category added successfully",
+      category,
+    });
+  } catch (error) {
+    console.error("Error adding 3rd level category:", error.message);
+    res.status(500).json({ error: "Server error" });
+  }
+};
+
+// Adding 4th level subcategory
+exports.addLavel4Category = async (req, res) => {
+  try {
+    const {
+      mainCategory,
+      subCategoryName,
+      subSubCategoryName,
+      lavel3CategoryName,
+      lavel4CategoryName,
+    } = req.body;
+
+    const category = await Category.findOne({ mainCategory });
+    if (!category) {
+      return res.status(404).json({ message: "Main category not found" });
+    }
+
+    const subCategory = category.subcategories.find(
+      (sub) => sub.name === subCategoryName
+    );
+    if (!subCategory) {
+      return res.status(404).json({ message: "Subcategory not found" });
+    }
+
+    const subSubCategory = subCategory.subsubcategories.find(
+      (subSub) => subSub.name === subSubCategoryName
+    );
+    if (!subSubCategory) {
+      return res.status(404).json({ message: "Subsubcategory not found" });
+    }
+
+    const lavel3Category = subSubCategory.lavel3CategorySchema.find(
+      (lavel3) => lavel3.name === lavel3CategoryName
+    );
+    if (!lavel3Category) {
+      return res.status(404).json({ message: "3rd level category not found" });
+    }
+
+    lavel3Category.lavel4CategorySchema.push({ name: lavel4CategoryName });
+    await category.save();
+    res.status(201).json({
+      message: "4th level category added successfully",
+      category,
+    });
+  } catch (error) {
+    console.error("Error adding 4th level category:", error.message);
+    res.status(500).json({ error: "Server error" });
+  }
+};
+
+exports.getCategories = async (req, res) => {
   try {
     const categories = await Category.find();
     res.status(200).json(categories);
@@ -86,7 +175,7 @@ const getCategories = async (req, res) => {
   }
 };
 
-const updateCategory = async (req, res) => {
+exports.updateCategory = async (req, res) => {
   try {
     const {
       mainCategory,
@@ -95,6 +184,10 @@ const updateCategory = async (req, res) => {
       newSubCategoryName,
       subSubCategoryName,
       newSubSubCategoryName,
+      lavel3CategoryName,
+      newLavel3CategoryName,
+      lavel4CategoryName,
+      newLavel4CategoryName,
     } = req.body;
 
     const category = await Category.findOne({ mainCategory });
@@ -117,10 +210,8 @@ const updateCategory = async (req, res) => {
     }
 
     if (subSubCategoryName && newSubSubCategoryName) {
-      const subCategory = category.subcategories.find((sub) =>
-        sub.subsubcategories.find(
-          (subSub) => subSub.name === subSubCategoryName
-        )
+      const subCategory = category.subcategories.find(
+        (sub) => sub.name === subCategoryName
       );
       if (!subCategory) {
         return res.status(404).json({ message: "Subcategory not found" });
@@ -134,21 +225,90 @@ const updateCategory = async (req, res) => {
       subSubCategory.name = newSubSubCategoryName;
     }
 
+    if (lavel3CategoryName && newLavel3CategoryName) {
+      const subCategory = category.subcategories.find(
+        (sub) => sub.name === subCategoryName
+      );
+      if (!subCategory) {
+        return res.status(404).json({ message: "Subcategory not found" });
+      }
+      const subSubCategory = subCategory.subsubcategories.find(
+        (subSub) => subSub.name === subSubCategoryName
+      );
+      if (!subSubCategory) {
+        return res.status(404).json({ message: "Subsubcategory not found" });
+      }
+      const lavel3Category = subSubCategory.lavel3CategorySchema.find(
+        (lavel3) => lavel3.name === lavel3CategoryName
+      );
+      if (!lavel3Category) {
+        return res
+          .status(404)
+          .json({ message: "3rd level category not found" });
+      }
+      lavel3Category.name = newLavel3CategoryName;
+    }
+
+    if (lavel4CategoryName && newLavel4CategoryName) {
+      const subCategory = category.subcategories.find(
+        (sub) => sub.name === subCategoryName
+      );
+      if (!subCategory) {
+        return res.status(404).json({ message: "Subcategory not found" });
+      }
+      const subSubCategory = subCategory.subsubcategories.find(
+        (subSub) => subSub.name === subSubCategoryName
+      );
+      if (!subSubCategory) {
+        return res.status(404).json({ message: "Subsubcategory not found" });
+      }
+      const lavel3Category = subSubCategory.lavel3CategorySchema.find(
+        (lavel3) => lavel3.name === lavel3CategoryName
+      );
+      if (!lavel3Category) {
+        return res
+          .status(404)
+          .json({ message: "3rd level category not found" });
+      }
+      const lavel4Category = lavel3Category.lavel4CategorySchema.find(
+        (lavel4) => lavel4.name === lavel4CategoryName
+      );
+      if (!lavel4Category) {
+        return res
+          .status(404)
+          .json({ message: "4th level category not found" });
+      }
+      lavel4Category.name = newLavel4CategoryName;
+    }
+
     await category.save();
-    res
-      .status(200)
-      .json({ message: "Category updated successfully", category });
+    res.status(200).json({
+      message: "Category updated successfully",
+      category,
+    });
   } catch (error) {
     console.error("Error updating category:", error.message);
     res.status(500).json({ error: "Server error" });
   }
 };
 
-const deleteCategory = async (req, res) => {
+exports.deleteCategory = async (req, res) => {
   try {
-    const { mainCategory, subCategoryName, subSubCategoryName } = req.body;
+    const {
+      mainCategory,
+      subCategoryName,
+      subSubCategoryName,
+      lavel3CategoryName,
+      lavel4CategoryName,
+    } = req.body;
 
-    if (mainCategory && !subCategoryName && !subSubCategoryName) {
+    if (
+      mainCategory &&
+      !subCategoryName &&
+      !subSubCategoryName &&
+      !lavel3CategoryName &&
+      !lavel4CategoryName
+    ) {
       await Category.deleteOne({ mainCategory });
       return res
         .status(200)
@@ -160,7 +320,12 @@ const deleteCategory = async (req, res) => {
       return res.status(404).json({ message: "Main category not found" });
     }
 
-    if (subCategoryName && !subSubCategoryName) {
+    if (
+      subCategoryName &&
+      !subSubCategoryName &&
+      !lavel3CategoryName &&
+      !lavel4CategoryName
+    ) {
       const subCategoryIndex = category.subcategories.findIndex(
         (sub) => sub.name === subCategoryName
       );
@@ -169,12 +334,18 @@ const deleteCategory = async (req, res) => {
       }
       category.subcategories.splice(subCategoryIndex, 1);
       await category.save();
-      return res
-        .status(200)
-        .json({ message: "Subcategory deleted successfully", category });
+      return res.status(200).json({
+        message: "Subcategory deleted successfully",
+        category,
+      });
     }
 
-    if (subCategoryName && subSubCategoryName) {
+    if (
+      subCategoryName &&
+      subSubCategoryName &&
+      !lavel3CategoryName &&
+      !lavel4CategoryName
+    ) {
       const subCategory = category.subcategories.find(
         (sub) => sub.name === subCategoryName
       );
@@ -189,9 +360,86 @@ const deleteCategory = async (req, res) => {
       }
       subCategory.subsubcategories.splice(subSubCategoryIndex, 1);
       await category.save();
-      return res
-        .status(200)
-        .json({ message: "Subsubcategory deleted successfully", category });
+      return res.status(200).json({
+        message: "Subsubcategory deleted successfully",
+        category,
+      });
+    }
+
+    if (
+      subCategoryName &&
+      subSubCategoryName &&
+      lavel3CategoryName &&
+      !lavel4CategoryName
+    ) {
+      const subCategory = category.subcategories.find(
+        (sub) => sub.name === subCategoryName
+      );
+      if (!subCategory) {
+        return res.status(404).json({ message: "Subcategory not found" });
+      }
+      const subSubCategory = subCategory.subsubcategories.find(
+        (subSub) => subSub.name === subSubCategoryName
+      );
+      if (!subSubCategory) {
+        return res.status(404).json({ message: "Subsubcategory not found" });
+      }
+      const lavel3CategoryIndex = subSubCategory.lavel3CategorySchema.findIndex(
+        (lavel3) => lavel3.name === lavel3CategoryName
+      );
+      if (lavel3CategoryIndex === -1) {
+        return res
+          .status(404)
+          .json({ message: "3rd level category not found" });
+      }
+      subSubCategory.lavel3CategorySchema.splice(lavel3CategoryIndex, 1);
+      await category.save();
+      return res.status(200).json({
+        message: "3rd level category deleted successfully",
+        category,
+      });
+    }
+
+    if (
+      subCategoryName &&
+      subSubCategoryName &&
+      lavel3CategoryName &&
+      lavel4CategoryName
+    ) {
+      const subCategory = category.subcategories.find(
+        (sub) => sub.name === subCategoryName
+      );
+      if (!subCategory) {
+        return res.status(404).json({ message: "Subcategory not found" });
+      }
+      const subSubCategory = subCategory.subsubcategories.find(
+        (subSub) => subSub.name === subSubCategoryName
+      );
+      if (!subSubCategory) {
+        return res.status(404).json({ message: "Subsubcategory not found" });
+      }
+      const lavel3Category = subSubCategory.lavel3CategorySchema.find(
+        (lavel3) => lavel3.name === lavel3CategoryName
+      );
+      if (!lavel3Category) {
+        return res
+          .status(404)
+          .json({ message: "3rd level category not found" });
+      }
+      const lavel4CategoryIndex = lavel3Category.lavel4CategorySchema.findIndex(
+        (lavel4) => lavel4.name === lavel4CategoryName
+      );
+      if (lavel4CategoryIndex === -1) {
+        return res
+          .status(404)
+          .json({ message: "4th level category not found" });
+      }
+      lavel3Category.lavel4CategorySchema.splice(lavel4CategoryIndex, 1);
+      await category.save();
+      return res.status(200).json({
+        message: "4th level category deleted successfully",
+        category,
+      });
     }
 
     res.status(400).json({ message: "Invalid delete request" });
@@ -199,13 +447,4 @@ const deleteCategory = async (req, res) => {
     console.error("Error deleting category:", error.message);
     res.status(500).json({ error: "Server error" });
   }
-};
-
-module.exports = {
-  createMainCategory,
-  addSubCategory,
-  addSubSubCategory,
-  getCategories,
-  updateCategory,
-  deleteCategory,
 };
